+++ Assets/Packages/GameSaver/Scripts/Core/SaveSettings.cs
@@ -7,8 +7,8 @@ namespace ThanhDV.GameSaver.Core
     [CreateAssetMenu(fileName = "SaveSettings", menuName = "GameSaver/SaveSettings", order = 0)]
     public class SaveSettings : ScriptableObject
     {
-        [Header("Storage Settings")]
-        [SerializeField] private bool createProfileOnFirstRun = true; public bool CreateProfileOnFirstRun => createProfileOnFirstRun;
+        [UnderlineHeader("Storage Settings")]
+        [SerializeField] private bool createProfileIfNull = true; public bool CreateProfileIfNull => createProfileIfNull;
         [SerializeField] private bool useEncryption = true; public bool UseEncryption => useEncryption;
 
         [SerializeField, Tooltip("If checked, each data module will be saved as a separate file.")]
@@ -16,7 +16,7 @@ namespace ThanhDV.GameSaver.Core
         [SerializeField, HideIf("saveAsSeparateFiles")] private string fileName = Constant.DEFAULT_FILE_NAME; public string FileName => fileName;
         [SerializeField, Tooltip("File extension (include leading dot).")] private string fileExtension = Constant.DEFAULT_FILE_SAVE_EXTENSION; public string FileExtension => fileExtension;
 
-        [Header("Auto Save")]
+        [UnderlineHeader("Auto Save")]
         [SerializeField] private bool enableAutoSave = true; public bool EnableAutoSave => enableAutoSave;
         [SerializeField, Tooltip("Second."), ShowIf("enableAutoSave")] private float autoSaveTime = 300f; public float AutoSaveTime => autoSaveTime;
     }