+++ Assets/Packages/GameSaver/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.0.2] - 2025-09-22
+- Fix Header Attribute
+- Update How to Use
+- Add method `WakeUp()` to Initialize via code.
+
 ## [1.0.1] - 2025-09-18
 - Fix bug on build game